backend/pybooking/app/requirements.txt
fastapi[standard]
python-dotenv
psycopg2-binary
uvicorn[standard]
google-genai
pillow